admin dashboard

+++ scribble/database/migrations/2024_05_25_081754_create_voucher_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('VoucherName');
             $table->date('StartDate');
             $table->date('EndDate');
-            $table->integer('Value');
+            $table->float('Value');
             $table->string('DiscountCategory');
             $table->timestamps();
         });
